[OV-213] feat/ redis config 설정 (#326)
