+++ delivery/build.gradle
@@ -38,6 +38,9 @@ dependencies {
 	// ✅ redis
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 
+	// ✅ redisson
+	implementation 'org.redisson:redisson:3.22.0'
+
 	//hibernate-type
 //	implementation group: 'com.vladmihalcea', name: 'hibernate-types-60', version: '2.21.0'
 	implementation 'io.hypersistence:hypersistence-utils-hibernate-63:3.9.5'
