+++ deliveryManager/build.gradle
@@ -35,6 +35,9 @@ dependencyManagement {
 
 dependencies {
 
+    // ✅ redisson
+    implementation 'org.redisson:redisson:3.22.0'
+
     //jpa
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
