+++ hub/build.gradle
@@ -91,12 +91,8 @@ dependencies {
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
-
-    // ✅redis
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-
     // ✅ redisson
-    implementation group: 'org.redisson', name: 'redisson', version: '3.45.0'
+    implementation 'org.redisson:redisson:3.22.0'
 }
 
 tasks.named('test') {
