+++ hub/src/main/java/com/owlexpress/hub/infrastructure/config/RedissonConfig.java
@@ -1,0 +1,52 @@
+package com.owlexpress.hub.infrastructure.config;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+import org.redisson.Redisson;
+import org.redisson.api.RedissonClient;
+import org.redisson.config.Config;
+import org.redisson.spring.cache.CacheConfig;
+import org.redisson.spring.cache.RedissonSpringCacheManager;
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.cache.CacheKeyPrefix;
+import org.springframework.data.redis.cache.RedisCacheConfiguration;
+import org.springframework.data.redis.cache.RedisCacheManager;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.serializer.RedisSerializer;
+
+import java.time.Duration;
+
+import static org.springframework.data.redis.serializer.RedisSerializationContext.SerializationPair.fromSerializer;
+
+@EnableCaching
+@Configuration
+public class RedissonConfig {
+
+    @Bean
+    public RedissonClient redissonClient() {
+        Config config = new Config();
+        config.useSingleServer()
+                .setAddress("redis://localhost:6378");
+        return Redisson.create(config);
+    }
+
+    @Bean
+    public CacheManager cacheManager(
+            RedissonClient redissonClient
+    ) {
+        Map<String, org.redisson.spring.cache.CacheConfig> config = new HashMap<>();
+
+        config.put("cacheName", new CacheConfig(
+                TimeUnit.HOURS.toMillis(1), // TTL
+                TimeUnit.MINUTES.toMillis(30) // Max idle time
+        ));
+
+        return new RedissonSpringCacheManager(redissonClient, config);
+    }
+}
